extracted the externally calling calendar/connectum code to a different file
